Move message input flow into `ReceivePath`
